Added tests of invalid template classes.
