+++ test/test_element.py
@@ -55,7 +55,7 @@ def test_set_value_eng(simple_element):
 
 def test_set_value_phys(simple_element, double_uc):
     simple_element._data_source_manager._uc['x'] = double_uc
-    simple_element.set_value('x', DUMMY_VALUE_2, handle=pytac.SP, 
+    simple_element.set_value('x', DUMMY_VALUE_2, handle=pytac.SP,
                              units=pytac.PHYS)
     # Conversion fron physics to engineering units
     simple_element.get_device('x').set_value.assert_called_with((DUMMY_VALUE_2 / 2))
@@ -65,7 +65,9 @@ def test_set_exceptions(simple_element, unit_uc):
     with pytest.raises(pytac.exceptions.FieldException):
         simple_element.set_value('unknown_field', 40.0, 'setpoint')
     with pytest.raises(pytac.exceptions.HandleException):
-        simple_element.set_value('y', 40.0, 'unknown_handle')
+        # No need for 'unknown_handle' as default handle is pytac.RB which will
+        # raise HandleException anyway as you cannot write on a readback pv.
+        simple_element.set_value('y', 40.0)
     with pytest.raises(pytac.exceptions.DeviceException):
         simple_element.set_value('y', 40.0, 'setpoint',
                                  data_source='unknown_data_source')
